refactor(runners): move VLARuntime from agents/ to models/openvla.py
VLARuntime duplicated the model loading logic already in openvla.py.
Now it lives alongside make_openvla_policy() — both are OpenVLA model
concerns, not agent orchestration concerns.

  models/openvla.py  — OpenVLARunner (training) + make_openvla_policy()
                       (fixed instruction) + VLARuntime (per-call instruction)
  agents/            — only orchestration: protocols, planner, planned

Deletes agents/vla.py. Same pattern as Gemma: model loading in models/,
planning logic in agents/.
