+++ src/odyssey/runners/agents/__init__.py
@@ -1,16 +1,16 @@
 """Multi-agent evaluation runtimes.
 
 Protocols:
-  * ``TextGenerator`` — chat messages → text (model-layer interface)
-  * ``PilotRuntime`` — image + instruction → action
-  * ``PlannerRuntime`` — task instruction → sub-instructions
+  * ``TextGenerator`` — chat messages -> text (model-layer interface)
+  * ``PilotRuntime`` — image + instruction -> action
+  * ``PlannerRuntime`` — task instruction -> sub-instructions
 
 Implementations:
-  * ``VLARuntime`` — OpenVLA (requires ``openvla`` extra)
   * ``LLMPlanner`` — planning logic, takes any TextGenerator
   * ``PlannedEvalRuntime`` — composes planner + pilot with phase transitions
 
-Model loading (``GemmaTextGenerator``, etc.) lives in ``runners/models/``.
+Model loading (``VLARuntime``, ``GemmaTextGenerator``) lives in
+``runners/models/``.
 """
 
 from odyssey.runners.agents.planned import (
@@ -30,12 +30,3 @@ __all__ = [
     "PlannerRuntime",
     "TextGenerator",
 ]
-
-
-def __getattr__(name: str) -> object:
-    """Lazy-import heavy implementations to avoid pulling in torch at import."""
-    if name == "VLARuntime":
-        from odyssey.runners.agents.vla import VLARuntime
-
-        return VLARuntime
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
